requirements.txt
Django==4.2.3
opencv-python==4.5.5.64
torch
torchvision
Pillow==9.5.0
trimesh==3.13.1
